:truck: Split test_grid into files
